[engine] Fix SQ bug: subtraction of `int`-s is always `int`
https://sonarcloud.io/project/issues?issues=AYTI3E9gdRT_9KYDIZ_l&open=AYTI3E9gdRT_9KYDIZ_l&id=vividus-framework_vividus
https://sonarcloud.io/organizations/vividus/rules?open=java%3AS2184&rule_key=java%3AS2184

+++ vividus-engine/src/main/java/org/vividus/log/LoggingTableTransformerMonitor.java
@@ -79,7 +79,7 @@ public class LoggingTableTransformerMonitor implements TableTransformerMonitor
                         table -> FORMATTING_TRANSFORMER.transform(table, tableParsers, tableProperties)
                 ))
                 .trim();
-        long numberOfTruncatedRows = totalNumberOrRows.get() - MAX_ROWS_LOGGING;
+        int numberOfTruncatedRows = totalNumberOrRows.get() - MAX_ROWS_LOGGING;
         if (numberOfTruncatedRows > 0)
         {
             return String.format("%s%s...%d row(s) truncated", resultingTable, rowSeparator, numberOfTruncatedRows);
